Create NEWscreensaver.py
The new and improved version of my screensaver includes all the elements I used before, just with small edits; but this time I added a new interesting shape. There is a burst in the top left corner that almost appears to flutter, and I attached it to all my old squares that bounced around with no reason.