Create NEWfrogclock.py
This addition to the Frog Clock makes it so the frog closes its eyes for a rest on the half hours. I tried to do this initially, however I was struggling to make the eyelids fit onto the frogs face, but after revisiting this clock, the frog can now get some rest and close its eyes. I think it's a cute edit and I'm really proud I got it to work with some assistance.