Create NEWsmothiebot.json
Simple edits were made to this bot, however I think it made it a lot more aesthetically pleasing. When I created my twitter bot, I added a bunch of cute emojis to give off a "vibe" to the bot,  so that inspired me to have each smoothie get a vibe with emojis also. I added an inspirational quote to top it all off and just make it a super uplifting and feel-good bot. I also added a lot more options to the milk section specifically, but I also added more fruit and toppings!